fix: type issues
